pyproject.toml
[tool.poetry]
name = "finance"
version = "0.1.0"
description = ""
authors = ["Akshay Bhardwaj"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = "^0.1.14"
jupyterlab = "^4.1.5"
unstructured = "^0.13.2"
libmagic = "^1.0"
python-magic = "^0.4.27"
faiss-cpu = "^1.8.0"
sentence-transformers = "^2.6.1"
python-dotenv = "^1.0.1"
openai = "^1.16.2"
streamlit = "^1.33.0"
langchain-openai = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
